Feat: delete post
1. 수정 articles/views.py, articles/detail.html, articles/urls.py
2. url에 delete를 추가하여 delete함수를 만들어 데이터를 불러와 삭제하기
   위해서 (form으로 post형식으로)

+++ Django Basic/articles/views.py
@@ -38,5 +38,8 @@ def detail(request, pk):
 
 def delete(request, pk): 
     article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('article:dinner')
+    
+    if request.method == 'POST':
+        article.delete()
+        
+    return redirect('articles:dinner')
